Park SoundEx (7-byte register tie); fix its int->short prototype
SoundEx (0x8004fc70, 512 bytes, 42 callers) parked NON_MATCHING: 7 differing
bytes, same length. The vertical-distance abs block colours `raw = locate->vy -
pp->vy` into $v1 where the target uses $v0 -- two freely interchangeable
caller-saved temps, no C-level lever. Same abs-value register-role class as
GetTargetDistance. Default ./Build check green, NON_MATCHING=SoundEx compiles,
symcheck clean.

item.h: SoundEx `int id` -> `short id` (PSX.SYM's prototype; needed for the
prologue sign-extension). The redundant local externs in PutMap.c and
ReqItemMakibishi.c updated to match. Verified safe: default build byte-identical.
gp-extern STAGE_SOUNDS_POINTER added (was compiling absolute).

Struct fields SoundEx proves: TCameraStatus.Mode @0x14 s32 (lw + compare to 1),
.DirectionRY @0x1A s16 (lh, signed) -- both re-confirm DoInfoViewProc's typedef;
Humanoid.rotate @0x3C -> SVECTOR.vy @0x2 signed (lh).

New cookbook rule: reuse a DYING call-result variable to carry a later
call-crossing value so they coalesce into one callee-saved register. The inverse
of the reduce-callee-saved levers -- for when the target's frame is LARGER than
yours (SoundEx: 61 diff lines collapsed to 5 by reusing `dist` for the volume).

+++ src/main.exe/PutMap.c
@@ -145,7 +145,7 @@ extern void *GsGetWorkBase(void);
 extern void GsSetWorkBase(void *workBase);
 extern void SetPolyXF4(POLY_XF4 *ply, short attrib);
 extern void GsSortSprite(GsSPRITE *sp, GsOT *ot, s32 pri);
-extern s16 SoundEx(VECTOR *loc, int id);
+extern s16 SoundEx(VECTOR *loc, short id);
 extern void FUN_8003d768(s32 x, s32 z, s32 *area);
 extern void AddXF4(void *ot, POLY_XF4 *ply);
