+++ src/main.exe/item.h
@@ -322,7 +322,7 @@ extern int ReqItemDrop(PARAM_ITEM_USE *p);
 extern int ReqItemStay(PARAM_ITEM_STAY *p);
 extern void SetBleed(VECTOR *pos, SVECTOR *vel, int a, int col);
 extern void SetSmoke(VECTOR *pos, SVECTOR *vel, short n, short time);
-extern s16 SoundEx(VECTOR *loc, int id);
+extern s16 SoundEx(VECTOR *loc, short id);
 extern void DeleteConflict(ModelType *m);
 extern void AdtMessageBox(char *fmt, ...);
 extern int rand(void);
